requirements.txt
numpy>=1.20.0
pandas>=1.3.0
scikit-learn>=1.0.0
fastapi>=0.75.0
uvicorn>=0.17.0
python-dotenv>=0.19.0
pydantic>=1.9.0
pytest>=7.0.0
requests>=2.27.0
matplotlib>=3.5.0
seaborn>=0.11.0
